Obsolete Coreversion.DEFAULT in the tests
We replace DEFAULT with the LATEST_STABLE, which currently refers to the
dotnet 6.0. It should be used in case test is not dependent on the
particular SDK version. In all other cases, dotnet version should be
specified explicitly, which makes coverage for the given framework
visible.

See [RIDER-76457](https://youtrack.jetbrains.com/issue/RIDER-76457).