Update test expectations in the template tests
This change related to the updated of the dotnet 7 SDK. Also it was
possible to use same gold file on mac and linux.
Linux fails to run tests on netcoreapp and has to be investigated
separately, this didn't work on netcoreapp2.1 before. Same test works
fine on Windows and macOS.

See [RIDER-81473](https://youtrack.jetbrains.com/issue/RIDER-81473).