[Rider] Inline doc squash, intermediate commit messages under cut!
[Rider] Revert accidental `.idea` directory edits

[Rider] Many minor fixes for Rider inline doc
Achieved correct inline doc presentation on the majority of R#/.cs files

[Rider] Fixes for C# inline doc backendless mode
Also made backend daemon stage friendlier to multifile source files

[Rider] Patch up the C# inline docs
Add a backend-driven fallback for the C# inline docs
and fix extra <br />s in C# inline docs

[Rider] RIDER-107629 Verify backend languages inlinedoc availability
Made F# work with the inline doc infra in some simple cases
(without touching any of the F# plugin code!)

For any further backend-only inline documentation implementations it
would be enough to either implement the XmlDocs in their PSI structure
in a way that is recognizable by `RiderInlineDocStageProcess` or to edit
`RiderInlineDocStageProcess` so it would recognize the structure that
is already in place.

Maybe we should also implement some kind of a DI-y way to add more
processing to the RiderInlineDocStageProcess for people who would
attempt a feat of making their custom .net-based language usable
in Rider via a plugin?

[Rider] Disabled the reader mode configurable for Rider, since it doesn't work yet anyway

[Rider] Completed (more or less) the inline documentation feature.
There is a way to add inline docs for backend languages that are not C#, find Egor Skrypnikov to find out how it might be done!

[Rider] Inline documentation, new and improved

GitOrigin-RevId: aef37ea87ea8f7c7fa999a245057980102d7adf4